+++ Source/Examples/ExampleLibrary/Series/BoxPlotSeriesExamples.cs
@@ -13,6 +13,8 @@ namespace ExampleLibrary
     using System.Collections.Generic;
     using System.Linq;
 
+    using ExampleLibrary.Utilities;
+
     using OxyPlot;
     using OxyPlot.Annotations;
     using OxyPlot.Axes;
@@ -29,10 +31,10 @@ namespace ExampleLibrary
             var model = new PlotModel { Title = string.Format("BoxPlot (n={0})", boxes), LegendPlacement = LegendPlacement.Outside };
 
             var s1 = new BoxPlotSeries
-                {
-                    Title = "BoxPlotSeries",
-                    BoxWidth = 0.3
-                };
+            {
+                Title = "BoxPlotSeries",
+                BoxWidth = 0.3
+            };
 
             var random = new Random(31);
             for (var i = 0; i < boxes; i++)
@@ -68,6 +70,12 @@ namespace ExampleLibrary
             model.Axes.Add(new LinearAxis { Position = AxisPosition.Left });
             model.Axes.Add(new LinearAxis { Position = AxisPosition.Bottom, MinimumPadding = 0.1, MaximumPadding = 0.1 });
             return model;
+        }
+
+        [Example("BoxPlot (transposed)")]
+        public static PlotModel BoxPlotTransposed()
+        {
+            return BoxPlot().Transpose();
         }
 
         /// <summary>