+++ Source/Examples/ExampleLibrary/Series/HistogramSeriesExamples.cs
@@ -11,6 +11,8 @@ namespace ExampleLibrary
 {
     using System;
 
+    using ExampleLibrary.Utilities;
+
     using OxyPlot;
     using OxyPlot.Axes;
     using OxyPlot.Series;
@@ -23,6 +25,57 @@ namespace ExampleLibrary
         public static PlotModel ExponentialDistribution()
         {
             return CreateExponentialDistribution();
+        }
+
+        [Example("Exponential Distribution (transposed)")]
+        public static PlotModel ExponentialDistributionTransposed()
+        {
+            return ExponentialDistribution().Transpose();
+        }
+
+        [Example("Label Placement")]
+        public static PlotModel HistogramLabelPlacement()
+        {
+            var model = new PlotModel { Title = "Label Placement" };
+
+            var s1 = new HistogramSeries { LabelPlacement = LabelPlacement.Base, LabelFormatString = "Base", StrokeThickness = 1, LabelMargin = 5 }; 
+            var s2 = new HistogramSeries { LabelPlacement = LabelPlacement.Inside, LabelFormatString = "Inside", StrokeThickness = 1, LabelMargin = 5 };
+            var s3 = new HistogramSeries { LabelPlacement = LabelPlacement.Middle, LabelFormatString = "Middle", StrokeThickness = 1, LabelMargin = 5 };
+            var s4 = new HistogramSeries { LabelPlacement = LabelPlacement.Outside, LabelFormatString = "Outside", StrokeThickness = 1, LabelMargin = 5 };
+
+            s1.Items.Add(new HistogramItem(1, 2, 4));
+            s1.Items.Add(new HistogramItem(2, 3, -4));
+            s2.Items.Add(new HistogramItem(3, 4, 2));
+            s2.Items.Add(new HistogramItem(4, 5, -2));
+            s3.Items.Add(new HistogramItem(5, 6, 3));
+            s3.Items.Add(new HistogramItem(6, 7, -3));
+            s4.Items.Add(new HistogramItem(7, 8, 1));
+            s4.Items.Add(new HistogramItem(8, 9, -1));
+
+            model.Series.Add(s1);
+            model.Series.Add(s2);
+            model.Series.Add(s3);
+            model.Series.Add(s4);
+
+            return model;
+        }
+
+        [Example("Label Placement (transposed)")]
+        public static PlotModel LabelPlacementTransposed()
+        {
+            return HistogramLabelPlacement().Transpose();
+        }
+
+        [Example("Label Placement (reversed Y Axis)")]
+        public static PlotModel LabelPlacementReversed()
+        {
+            return HistogramLabelPlacement().ReverseYAxis();
+        }
+
+        [Example("Label Placement (reversed Y Axis, transposed)")]
+        public static PlotModel LabelPlacementReversedTransposed()
+        {
+            return LabelPlacementReversed().Transpose();
         }
 
         [Example("Custom Bins")]
@@ -43,7 +96,7 @@ namespace ExampleLibrary
             model.Axes.Add(new LinearAxis { Position = AxisPosition.Left, Title = "Frequency" });
             model.Axes.Add(new LinearAxis { Position = AxisPosition.Bottom, Title = "x" });
 
-            Random rnd = new Random();
+            Random rnd = new Random(1);
 
             HistogramSeries chs = new HistogramSeries();
             chs.Items.AddRange(HistogramHelpers.Collect(SampleExps(rnd, 1.0, n), 0, 5, 15, true));
@@ -76,8 +129,6 @@ namespace ExampleLibrary
             var model = new PlotModel { Title = "Disconnected Bins" };
             model.Axes.Add(new LinearAxis { Position = AxisPosition.Left, Title = "Representation" });
             model.Axes.Add(new LinearAxis { Position = AxisPosition.Bottom, Title = "x" });
-
-            Random rnd = new Random();
 
             HistogramSeries chs = new HistogramSeries();
             chs.Items.AddRange(new[] { new HistogramItem(0, 0.5, 10), new HistogramItem(0.75, 1.0, 10) });