+++ Source/Examples/ExampleLibrary/Series/StemSeriesExamples.cs
@@ -11,7 +11,10 @@ namespace ExampleLibrary
 {
     using System;
 
+    using ExampleLibrary.Utilities;
+
     using OxyPlot;
+    using OxyPlot.Axes;
     using OxyPlot.Series;
 
     /// <summary>
@@ -32,6 +35,12 @@ namespace ExampleLibrary
                              MarkerFill = OxyColors.SkyBlue,
                              MarkerStrokeThickness = 1.5
                          });
+        }
+
+        [Example("StemSeries (transposed)")]
+        public static PlotModel StemSeriesTransposed()
+        {
+            return StemSeries().Transpose();
         }
 
         /// <summary>