+++ Source/OxyPlot/Series/FinancialSeries/CandleStickSeries.cs
@@ -92,11 +92,9 @@ namespace OxyPlot.Series
             var clippingRect = this.GetClippingRect();
             var dashArray = this.LineStyle.GetDashArray();
 
-            var datacandlewidth = (this.CandleWidth > 0) ? this.CandleWidth : this.minDx * 0.80;
-            var candlewidth = 
-                this.XAxis.Transform(items[0].X + datacandlewidth) -
-                this.XAxis.Transform(items[0].X); 
-
+            var dataCandlewidth = (this.CandleWidth > 0) ? this.CandleWidth : this.minDx * 0.80;
+            var halfDataCandlewidth = .5 * dataCandlewidth;
+            
             // colors
             var fillUp = this.GetSelectableFillColor(this.IncreasingColor);
             var fillDown = this.GetSelectableFillColor(this.DecreasingColor);
@@ -129,18 +127,15 @@ namespace OxyPlot.Series
 
                 var high = this.Transform(bar.X, bar.High);
                 var low = this.Transform(bar.X, bar.Low);
-
-                var open = this.Transform(bar.X, bar.Open);
-                var close = this.Transform(bar.X, bar.Close);
-                var max = new ScreenPoint(open.X, Math.Max(open.Y, close.Y));
-                var min = new ScreenPoint(open.X, Math.Min(open.Y, close.Y));
+                var max = this.Transform(bar.X, Math.Max(bar.Open, bar.Close));
+                var min = this.Transform(bar.X, Math.Min(bar.Open, bar.Close));
 
                 if (this.StrokeThickness > 0 && this.LineStyle != LineStyle.None)
                 {
                     // Upper extent
                     rc.DrawClippedLine(
                         clippingRect,
-                        new[] { high, min },
+                        new[] { high, max },
                         0,
                         lineColor,
                         this.StrokeThickness,
@@ -151,7 +146,7 @@ namespace OxyPlot.Series
                     // Lower extent
                     rc.DrawClippedLine(
                         clippingRect,
-                        new[] { max, low },
+                        new[] { min, low },
                         0,
                         lineColor,
                         this.StrokeThickness,
@@ -160,24 +155,10 @@ namespace OxyPlot.Series
                         true);
                 }
 
-                // Body
-                var openLeft = open + new ScreenVector(-candlewidth * 0.5, 0);
-
-                if (max.Y - min.Y < 1.0)
-                {
-                    var leftPoint = new ScreenPoint(openLeft.X - this.StrokeThickness, min.Y);
-                    var rightPoint = new ScreenPoint(openLeft.X + this.StrokeThickness + candlewidth, min.Y);
-                    rc.DrawClippedLine(clippingRect, new[] { leftPoint, rightPoint }, leftPoint.DistanceToSquared(rightPoint), lineColor, this.StrokeThickness, null, LineJoin.Miter, true);
-
-                    leftPoint = new ScreenPoint(openLeft.X - this.StrokeThickness, max.Y);
-                    rightPoint = new ScreenPoint(openLeft.X + this.StrokeThickness + candlewidth, max.Y);
-                    rc.DrawClippedLine(clippingRect, new[] { leftPoint, rightPoint }, leftPoint.DistanceToSquared(rightPoint), lineColor, this.StrokeThickness, null, LineJoin.Miter, true);
-                }
-                else
-                {
-                    var rect = new OxyRect(openLeft.X, min.Y, candlewidth, max.Y - min.Y);
-                    rc.DrawClippedRectangleAsPolygon(clippingRect, rect, fillColor, lineColor, this.StrokeThickness);
-                }
+                var p1 = this.Transform(bar.X - halfDataCandlewidth, bar.Open);
+                var p2 = this.Transform(bar.X + halfDataCandlewidth, bar.Close);
+                var rect = new OxyRect(p1, p2);
+                rc.DrawClippedRectangleAsPolygon(clippingRect, rect, fillColor, lineColor, this.StrokeThickness);
             }
         }
 
@@ -193,11 +174,7 @@ namespace OxyPlot.Series
             double yclose = legendBox.Top + ((legendBox.Bottom - legendBox.Top) * 0.3);
             double[] dashArray = this.LineStyle.GetDashArray();
 
-            var datacandlewidth = (this.CandleWidth > 0) ? this.CandleWidth : this.minDx * 0.80;
-
-            var candlewidth = Math.Min(
-                legendBox.Width,
-                this.XAxis.Transform(this.Items[0].X + datacandlewidth) - this.XAxis.Transform(this.Items[0].X));
+            var candlewidth = legendBox.Width * 0.75;
 
             if (this.StrokeThickness > 0 && this.LineStyle != LineStyle.None)
             {