+++ Source/OxyPlot/Series/FinancialSeries/VolumeSeries.cs
@@ -11,6 +11,7 @@ namespace OxyPlot.Series
 {
     using System;
     using System.Collections.Generic;
+    using System.Diagnostics;
 
     /// <summary>
     /// Represents a dual view (candlestick + volume) series for OHLCV bars
@@ -204,12 +205,10 @@ namespace OxyPlot.Series
 
             this.VerifyAxes();
 
-            var clipping = this.GetClippingRect();
+            var clippingRect = this.GetClippingRect();
 
             var datacandlewidth = (this.BarWidth > 0) ? this.BarWidth : this.minDx * 0.80;
-            var candlewidth =
-                this.XAxis.Transform(items[0].X + datacandlewidth) -
-                this.XAxis.Transform(items[0].X) - this.StrokeThickness;
+            var halfDataCandleWidth = datacandlewidth * 0.5;
 
             // colors
             var fillUp = this.GetSelectableFillColor(this.PositiveColor);
@@ -242,67 +241,78 @@ namespace OxyPlot.Series
                     continue;
                 }
 
-                var leftX = this.XAxis.Transform(bar.X) - (this.BarWidth / 2.0);
-                var y0 = this.YAxis.Transform(0);
+                var p1 = this.Transform(bar.X - halfDataCandleWidth, 0);
 
                 switch (this.VolumeStyle)
                 {
                     case VolumeStyle.Combined:
                         {
-                            var adj = this.YAxis.Transform(Math.Abs(bar.BuyVolume - bar.SellVolume));
+                            var p2 = this.Transform(bar.X + halfDataCandleWidth, Math.Abs(bar.BuyVolume - bar.SellVolume));
                             var fillcolor = (bar.BuyVolume > bar.SellVolume) ? barfillUp : barfillDown;
                             var linecolor = (bar.BuyVolume > bar.SellVolume) ? lineUp : lineDown;
-                            var rect1 = new OxyRect(leftX, adj, candlewidth, Math.Abs(adj - y0));
-                            rc.DrawClippedRectangleAsPolygon(clipping, rect1, fillcolor, linecolor, this.StrokeThickness);
+                            var rect1 = new OxyRect(p1, p2);
+                            rc.DrawClippedRectangleAsPolygon(clippingRect, rect1, fillcolor, linecolor, this.StrokeThickness);
                         }
 
                         break;
 
                     case VolumeStyle.PositiveNegative:
                         {
-                            var buyY = this.YAxis.Transform(bar.BuyVolume);
-                            var sellY = this.YAxis.Transform(-bar.SellVolume);
-                            var rect1 = new OxyRect(leftX, buyY, candlewidth, Math.Abs(buyY - y0));
-                            rc.DrawClippedRectangleAsPolygon(clipping, rect1, fillUp, lineUp, this.StrokeThickness);
-                            var rect2 = new OxyRect(leftX, y0, candlewidth, Math.Abs(sellY - y0));
-                            rc.DrawClippedRectangleAsPolygon(clipping, rect2, fillDown, lineDown, this.StrokeThickness);
+                            var p2Buy = this.Transform(bar.X + halfDataCandleWidth, bar.BuyVolume);
+                            var p2Bell = this.Transform(bar.X + halfDataCandleWidth, -bar.SellVolume);
+
+                            var rectBuy = new OxyRect(p1, p2Buy);
+                            var rectSell = new OxyRect(p1, p2Bell);
+
+                            rc.DrawClippedRectangleAsPolygon(clippingRect, rectBuy, fillUp, lineUp, this.StrokeThickness);
+                            rc.DrawClippedRectangleAsPolygon(clippingRect, rectSell, fillDown, lineDown, this.StrokeThickness);
                         }
 
                         break;
 
                     case VolumeStyle.Stacked:
-                        if (bar.BuyVolume > bar.SellVolume)
                         {
-                            var buyY = this.YAxis.Transform(bar.BuyVolume);
-                            var sellY = this.YAxis.Transform(bar.SellVolume);
-                            var dyoffset = sellY - y0;
-                            var rect2 = new OxyRect(leftX, sellY, candlewidth, Math.Abs(sellY - y0));
-                            rc.DrawClippedRectangleAsPolygon(clipping, rect2, fillDown, lineDown, this.StrokeThickness);
-                            var rect1 = new OxyRect(leftX, buyY + dyoffset, candlewidth, Math.Abs(buyY - y0));
-                            rc.DrawClippedRectangleAsPolygon(clipping, rect1, fillUp, lineUp, this.StrokeThickness);
-                        }
-                        else
-                        {
-                            var buyY = this.YAxis.Transform(bar.BuyVolume);
-                            var sellY = this.YAxis.Transform(bar.SellVolume);
-                            var dyoffset = buyY - y0;
-                            var rect1 = new OxyRect(leftX, buyY, candlewidth, Math.Abs(buyY - y0));
-                            rc.DrawClippedRectangleAsPolygon(clipping, rect1, fillUp, lineUp, this.StrokeThickness);
-                            var rect2 = new OxyRect(leftX, sellY + dyoffset, candlewidth, Math.Abs(sellY - y0));
-                            rc.DrawClippedRectangleAsPolygon(clipping, rect2, fillDown, lineDown, this.StrokeThickness);
-                        }
+                            var p2Buy = this.Transform(bar.X + halfDataCandleWidth, bar.BuyVolume);
+                            var p2Sell = this.Transform(bar.X + halfDataCandleWidth, bar.SellVolume);
+                            var pBoth = this.Transform(bar.X - halfDataCandleWidth, bar.BuyVolume + bar.SellVolume);
 
+                            OxyRect rectBuy;
+                            OxyRect rectSell;
+
+                            if (bar.BuyVolume > bar.SellVolume)
+                            {
+                                rectSell = new OxyRect(p1, p2Sell);
+                                rectBuy = new OxyRect(p2Sell, pBoth);
+                            }
+                            else
+                            {
+                                rectBuy = new OxyRect(p1, p2Buy);
+                                rectSell = new OxyRect(p2Buy, pBoth);
+                            }
+
+                            rc.DrawClippedRectangleAsPolygon(clippingRect, rectBuy, fillUp, lineUp, this.StrokeThickness);
+                            rc.DrawClippedRectangleAsPolygon(clippingRect, rectSell, fillDown, lineDown, this.StrokeThickness);
+
+                            break;
+                        }
+                    case VolumeStyle.None:
                         break;
+                    default:
+                        throw new ArgumentOutOfRangeException();
                 }
             }
 
             if (this.InterceptStrokeThickness > 0 && this.InterceptLineStyle != LineStyle.None)
             {
                 // draw volume y=0 line
-                var intercept = this.YAxis.Transform(0);
+                var p1 = this.InverseTransform(clippingRect.BottomLeft);
+                var p2 = this.InverseTransform(clippingRect.TopRight);
+                var lineA = this.Transform(p1.X, 0);
+                var lineB = this.Transform(p2.X, 0);
+
                 rc.DrawClippedLine(
-                    clipping,
-                    new[] { new ScreenPoint(clipping.Left, intercept), new ScreenPoint(clipping.Right, intercept) },
+                    clippingRect,
+                    new[] { lineA, lineB },
                     0,
                     this.InterceptColor,
                     this.InterceptStrokeThickness,
@@ -324,14 +334,10 @@ namespace OxyPlot.Series
             double yclose = legendBox.Top + ((legendBox.Bottom - legendBox.Top) * 0.3);
             double[] dashArray = LineStyle.Solid.GetDashArray();
 
-            var datacandlewidth = (this.BarWidth > 0) ? this.BarWidth : this.minDx * 0.80;
-
             var fillUp = this.GetSelectableFillColor(this.PositiveColor);
             var lineUp = this.GetSelectableColor(this.PositiveColor.ChangeIntensity(this.StrokeIntensity));
 
-            var candlewidth =
-                this.XAxis.Transform(this.data[0].X + datacandlewidth) -
-                this.XAxis.Transform(this.data[0].X);
+            var candlewidth = legendBox.Width * 0.75;
 
             if (this.StrokeThickness > 0)
             {